Fix run_sad
This problem is related to this issue: https://github.com/yarnpkg/yarn/issues/685

The problem is that yarn copies the entire folder from `./tripetto-block-text/` to `./app/node_modules/tripetto-block-text`. This also includes the `node_modules` folder in `./tripetto-block-text/node_modules`. This folder also contains the `tripetto` package. So now, the text block will use this locale instance of Tripetto to register the block instead of the one used by the app.